[*]Fixes a calculation bug and replaces unsigned Long Long variables with int variables where they are no longer needed due to a new means to hedge against national culture overloads. [*]Fixes a crash bug that would take place when the Complex Philosophical trait is selected.  Related to No Unit type for GP assignment not being fully safe for traits.  Checked the whole stream of code on this now and it should now be good.  When ALL GP pts earned are not of a unit type and you've reached the threshold for GP birth, you will continue to attempt to birth a GP every round until you finally get a colored GP point earned.  Even one will make the birthing take place finally.  This should be a fairly rare scenario.
git-svn-id: svn://svn.code.sf.net/p/caveman2cosmos/code/trunk@10571 8bbd16b5-4c62-4656-ae41-5efa6c748c97

Former-commit-id: e8900102353e411ca63e0f56bc3329a411dd1e8c

+++ Sources/CvGameTextMgr.cpp
@@ -36805,12 +36805,12 @@ void CvGameTextMgr::parsePlayerTraits(CvWStringBuffer &szBuffer, PlayerTypes ePl
 	if (GC.getGameINLINE().isOption(GAMEOPTION_LEADERHEAD_LEVELUPS))
 	{
 		int iLeaderLevel = GET_PLAYER(ePlayer).getLeaderHeadLevel();
-		unsigned long long iNationalCulture = GET_PLAYER(ePlayer).getCulture();
+		int iNationalCulture = GET_PLAYER(ePlayer).getCulture();
 		int iGreaterCulture = GET_PLAYER(ePlayer).getGreaterCulture();
 		int iNextGreaterCulture = 0;
-		unsigned long long iNextLevelup = GET_PLAYER(ePlayer).getLeaderLevelupNextCultureTotal(iNextGreaterCulture);
+		int iNextLevelup = GET_PLAYER(ePlayer).getLeaderLevelupNextCultureTotal(iNextGreaterCulture);
 		int iNextGreaterCultureRemaining = 0;
-		unsigned long long iRemaining = GET_PLAYER(ePlayer).getLeaderLevelupCultureToEarn(iNextGreaterCultureRemaining);
+		int iRemaining = GET_PLAYER(ePlayer).getLeaderLevelupCultureToEarn(iNextGreaterCultureRemaining);
 		szBuffer.append(NEWLINE);
 		szBuffer.append(gDLL->getText("TXT_KEY_LEADER_LEVEL", iLeaderLevel));
 		szBuffer.append(NEWLINE);
